pyproject.toml
[project]
name = "lab01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.40.0",
    "asyncio>=4.0.0",
    "jupyter>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
